pyproject.toml
[tool.poetry]
name = "gpt4-snippies"
version = "0.1.1"
description = "code snippets generated by gpt4"
authors = ["Robbie Capps <rocapp@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt4_snippies"}]

[tool.poetry.dependencies]
python = "^3.11"
PyGObject-stubs = ">=2.10.0"
python-dotenv = ">=1.0.1"
pygobject = "^3.46.0"
requests = "^2.31.0"
openapi3-parser = "^1.1.16"

[tool.poetry.group.development.dependencies]
ipykernel = "^6.29.2"
pickleshare = "^0.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
